Move buildMatcher and filterFilesForRule (with their picomatch/relative deps) out of runner.ts into src/rule-files.ts. Pure relocation, no behaviour change; runner.ts was at the line cap and include/exclude matching is a distinct concern.
